AI_Models/requirements.txt
Flask
pandas
numpy
scipy
python-dotenv
nixtla